REFAC(day_09): attach the gtest by bazel

+++ day_09/task_1/mirage_maintenance.cpp
@@ -1,11 +1,11 @@
 #include "mirage_maintenance.hpp"
-#include <cassert>
 #include <algorithm>
+#include <cassert>
 
 void calcucale_one_step_difference(Report& report, const int depth)
 {
     assert(report.size() > 0U);
-    assert(report.size() > depth);
+    assert(static_cast<int>(report.size()) > depth);
 
     auto value_to_save = report.front();
     for(std::size_t i = 0; i < report.size() - depth; ++i)
@@ -18,10 +18,12 @@ void calcucale_one_step_difference(Report& report, const int depth)
 void calcucale_final_values(Report& report)
 {
     bool are_zeroes_reached{false};
-    for(int i = 1; i < report.size() && !are_zeroes_reached; ++i)
+    for(int i = 1; static_cast<std::size_t>(i) < report.size() && !are_zeroes_reached; ++i)
     {
         calcucale_one_step_difference(report, i);
-        are_zeroes_reached = std::all_of(std::begin(report), std::prev(std::end(report), i), [](int value){return value == 0;});
+        are_zeroes_reached =
+            std::all_of(std::begin(report), std::prev(std::end(report), i),
+                        [](int value) { return value == 0; });
     }
 }
 
@@ -29,7 +31,7 @@ void calculate_extrapolated_values(Report& report)
 {
     calcucale_final_values(report);
 
-    for(int i = 1; i < report.size(); i++)
+    for(int i = 1; static_cast<std::size_t>(i) < report.size(); i++)
     {
         report[i] = report[i] - report[i - 1];
     }
